+++ app/county/county_routes.py
@@ -1,7 +1,9 @@
-from fastapi import APIRouter, Depends, HTTPException
+from fastapi import APIRouter, Depends, HTTPException, status
 
 from app.auth import get_current_user
-from app.utility.vice_county.vc_checker import VcChecker, NoVcException
+from app.utility.vice_county.vc_checker import (
+    VcChecker, NoVcFoundWarning, NoVcAllocation
+)
 
 
 router = APIRouter()
@@ -46,6 +48,11 @@ async def read_county_by_gridref(gridref: str):
             "code": code,
             "name": VcChecker.get_name_from_code(code)
         }
-    except NoVcException as e:
+    except NoVcFoundWarning as e:
         raise HTTPException(
-            status_code=404, detail=str(e))
+            status_code=status.HTTP_404_NOT_FOUND, detail=str(e))
+    except NoVcAllocation:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Only British grid references are supported."
+        )