+++ test/conftest.py
@@ -10,15 +10,23 @@ from app.main import app
 from app.settings_env import get_env_settings
 from app.user.user_repo import UserRepo
 
-from .mocks import mock_env_settings, mock_create_db, mock_settings
+from .mocks import mock_env_settings, mock_env_tolerant_settings, mock_create_db, mock_settings
 
 
 @pytest.fixture(name="env")
 def env_fixture():
-    """Fixture which supplies the environment settings.
+    """Fixture which supplies the environment settings without rule tolerance.
 
     Use this in tests which do not use the client-fixture."""
     return mock_env_settings()
+
+
+@pytest.fixture(name="env_tolerant")
+def env_tolerant_fixture():
+    """Fixture which supplies the environment settings with rule tolerance.
+
+    Use this in tests which do not use the client-fixture."""
+    return mock_env_tolerant_settings()
 
 
 @pytest.fixture(name="engine")